requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
html2text>=2020.1.16
google-genai>=1.0.0
requests>=2.31.0
tzdata>=2024.1
